Disallow error-prone use of control node fake, add to tests instead
